# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Reservation System API",
        "description": "API REST para el sistema de reservas de hotel. Incluye gestión de amenities, tipos de habitación y habitaciones físicas.",
        "contact": {
            "name": "API Support",
            "email": "adiazy@uamv.edu.ni"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Obtener lista de huéspedes",
                "description": "Lista de huéspedes\n\nBúsqueda simple por query param q:\n- Busca en: first_name, last_name, email, document_number\n\nFiltros opcionales:\n- ?country_code=NI",
                "operationId": "56cc8fe33a100159d774722901ae37e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Búsqueda por nombre, email o número de documento",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "juan"
                        }
                    },
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Filtrar por código de país (2 caracteres)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 2,
                            "example": "NI"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de huéspedes obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Crear un nuevo huésped",
                "description": "Crear un nuevo huésped",
                "operationId": "e792b91966348e6bb3439a3d24883e1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 120,
                                        "example": "juan.perez@example.com",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "example": "+50588888888",
                                        "nullable": true
                                    },
                                    "document_type": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "example": "PASSPORT",
                                        "nullable": true
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "maxLength": 64,
                                        "example": "A1234567",
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "NI",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Cliente frecuente.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Huésped creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Obtener un huésped por ID",
                "description": "Obtener un huésped específico",
                "operationId": "4267610a6a43088364913e1b6071b25d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped obtenido exitosamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Actualizar un huésped",
                "description": "Actualizar un huésped",
                "operationId": "af858a767bd4c0c6e7fd978a27775871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 120,
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "nullable": true
                                    },
                                    "document_type": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "nullable": true
                                    },
                                    "document_number": {
                                        "type": "string",
                                        "maxLength": 64,
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Huésped actualizado exitosamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Eliminar un huésped",
                "description": "Eliminar un huésped\n\nOpción simple: Eliminación física (delete()).\nNota: En proyectos reales, a veces se evita borrar huéspedes por historial de reservas.\nAlternativa: Podrías usar soft deletes o marcar como inactivo.",
                "operationId": "627107a5c10ad7a3877e6ca545859ca8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del huésped",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Huésped eliminado exitosamente"
                    },
                    "404": {
                        "description": "Huésped no encontrado"
                    }
                }
            }
        },
...
    },
...
}